+++ activesupport/lib/active_support/actionable_error.rb
@@ -9,7 +9,15 @@ module ActiveSupport
   module ActionableError
     extend Concern
 
-    class NonActionable < StandardError; end
+    class NonActionable < StandardError; end # :nodoc:
+
+    Trigger = Struct.new(:actionable, :condition) do # :nodoc:
+      def act_on(error)
+        raise actionable if condition&.call(error)
+      end
+    end
+
+    mattr_accessor :triggers, default: Hash.new { |h, k| h[k] = [] } # :nodoc:
 
     included do
       class_attribute :_actions, default: {}
@@ -30,6 +38,11 @@ module ActiveSupport
       raise NonActionable, "Cannot find action \"#{name}\""
     end
 
+    def self.raise_if_triggered_by(error) # :nodoc:
+      triggers[error.class].each { |trigger| trigger.act_on(error) }
+      raise_if_triggered_by(error.cause) if error.cause
+    end
+
     module ClassMethods
       # Defines an action that can resolve the error.
       #
@@ -42,6 +55,23 @@ module ActiveSupport
       #   end
       def action(name, &block)
         _actions[name] = block
+      end
+
+      # Trigger the current actionable error from a pre-existing if the
+      # condition, given as a block, returns a truthy value.
+      #
+      #   class SetupError < Error
+      #     include ActiveSupport::ActionableError
+      #
+      #     trigger on: ActiveRecord::RecordInvalid, if: -> error do
+      #       error.message.match?(InboundEmail.table_name)
+      #     end
+      #   end
+      def trigger(options)
+        error = options.fetch(:on) { raise ArgumentError, "missing keyword: on" }
+        condition = options.fetch(:if) { raise ArgumentError, "missing keyword: if" }
+
+        ActiveSupport::ActionableError.triggers[error] << Trigger.new(self, condition)
       end
     end
   end